Extracted PaymentForm logic to a custom hook, usePaymentForm
